+++ CHANGELOG.rst
@@ -32,6 +32,7 @@ Added
 - Example function in ``examples/c_tebd.py`` how to run TEBD with a model originally having next-nearest neighbors.
 - :meth:`~tenpy.networks.mps.MPS.increase_L` to allow increasing the unit cell of an MPS.
 - Argument `insert_all_id` for :meth:`tenpy.networks.mpo.MPOGraph.from_terms` and :meth:`~tenpy.networks.mpo.MPOGraph.from_term_list`
+- Additional option ``order='folded'`` for the :class:`~tenpy.models.lattice.Chain`.
 
 Fixed
 ^^^^^